% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest neighbors cross-validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of the training data.}

\item{k_nn}{A integer representing the number of neighbors.}

\item{k_cv}{A integer representing the number of folds.}
}
\value{
A list with objects \code{class}, which is a vector that includes
the predicted class for all observations, and \code{cv_err}, which is
a numeric with the cross-validation misclassification error.
}
\description{
This function performs k-nearest neighbors cross-validation test to predict class using given covariates.
}
\examples{
peng_clean <- na.omit(my_penguins)
my_knn_cv(peng_clean[, 3:6], peng_clean$species, 1, 5)

}
\keyword{k-nearest}
\keyword{neighbors}
