% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear regression model}
\usage{
my_lm(my_formula, my_data)
}
\arguments{
\item{my_formula}{A \code{formula} class object.}

\item{my_data}{Input data frame.}
}
\value{
Return a table with rows for each coefficient and columns for
\code{Estimate}, \code{Std. Error}, \code{t value}, and \code{p value}.
}
\description{
This function fits a linear regression model to given data.
}
\examples{
grades_data <- read.csv("https://www.openintro.org/data/csv/gpa.csv")
my_lm(gpa ~ studyweek, grades_data)

}
\keyword{lm}
