% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{One sample t test}
\usage{
my_t.test(x, alternative = c("two.sided", "less", "greater"), mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis,
which only accept \code{"two.sided"}, \code{"less"}, or \code{"greater"}.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements \code{test_stat} representing the numeric test statistic,
\code{df} representing the degrees of freedom, \code{alternative} representing
the value of the parameter \code{alternative}, and \code{p_val} representing
the numeric p-value.
}
\description{
This function evaluates if the value of the given numeric vector is significantly different from the null value (mu).
}
\examples{
helium_data <- read.csv("https://www.openintro.org/data/csv/helium.csv")
my_t.test(x = helium_data$helium, alternative = "two.sided", mu = 20)
my_t.test(x = helium_data$helium, alternative = "greater", mu = 20)

}
\keyword{t.test}
